Uploaded example.html
This file is a intended to show how CSS can be used to build a simple but effective dialog box.  JS is used to set the content of the message box and to show the background and the foreground elements.

The background element is intended to "white out" the web page behind it to prevent the user from clicking on things while the dialog is visible.  The foreground element is designed to automatically center the message box in front of the background.